Apply Shape Key to Basis Blender add-on
It has been part of CATS for a few years, but CATS is close to abandoned
at this point.

Separating the Operator out allows it to be installed individually and
may help prepare it for the Blender Extensions platform.

The code has been reorganised and many of the comments have been altered
to better match Blender's code guidelines.

In particular, reference to 'Basis' is now only user facing, because
'Basis' is simply the default name of the Reference Key.