Much faster Shape Key coordinates access using pointers
Shape Key coordinates have internal Blender issues, to do with their
differing typing for meshes and curves, that make them much slower to
access than most properties.

Shape Key access will now access the memory of the coordinates directly,
which is much faster.

The performance improvement is better the larger the number of vertices
in the mesh.

For a mesh with 90,000 vertices, it is about 40 times faster.